Fix importing problem with code duplication, fix visibility check
